Added welcome view moved files, and edited UI
