+++ ShufMath/Views/GamePlay/GameContainer.swift
@@ -11,21 +11,31 @@ struct GameContainer: View {
     @ObservedObject var viewModel: GameViewModel
     
     var body: some View {
-        switch viewModel.gameState {
-            case .notStarted:
-                StartingScreen(viewModel: viewModel)
-            case .inProgress:
-                MainGameView(viewModel: viewModel)
-            case .finished:
-                Color.clear
-                    .sheet(isPresented: $viewModel.isGameOver, onDismiss: {viewModel.playAgain()}) {
-                        GameOverView(viewModel: viewModel)
-                    }
+        
+        ZStack {
+            BackGroundView()
+            switch viewModel.gameState {
+                case .notStarted:
+                    StartingScreen(viewModel: viewModel)
+                    .transition(.move(edge: .leading))
+                case .inProgress:
+                        MainGameView(viewModel: viewModel)
+                    .transition(.move(edge: .trailing))
+                case .finished:
+                    Color.clear
+                        .sheet(isPresented: $viewModel.isGameOver, onDismiss: {viewModel.playAgain()}) {
+                            GameOverView(viewModel: viewModel)
+                        }
+            }
         }
+        .navigationBarBackButtonHidden(true) 
+
         GameAlert(viewModel: viewModel)
         // Changes isGameOver boolean for alerts based on the gameState
         .onChange(of: viewModel.gameState) {
             viewModel.isGameOver = viewModel.gameState == .finished
         }
+        .animation(.easeInOut(duration: 0.5), value: viewModel.gameState)
     }
+
 }
