+++ ShufMath/Views/Setup/Settings/GameModeSelector.swift
@@ -28,6 +28,5 @@ struct GameModeSelector: View {
                 action: {viewModel.setGameMode(.mixed)}
             )
         }
-
     }
 }
