+++ ShufMath/Views/Setup/Settings/GameSetup.swift
@@ -20,9 +20,7 @@ struct GameSetupView: View {
         VStack(spacing: 25){
             Section() {
                 Text("Game Setup")
-                    .font(.headline)
-                    .fontWeight(.bold)
-                    .foregroundColor(.white)
+                    .titleView()
                     .accessibilityLabel("Game Setup")
                     .accessibilityHint("This page will set up your game")
                 
@@ -36,6 +34,9 @@ struct GameSetupView: View {
                 )
             }
         }
-//        .padding()
     }
+}
+
+#Preview {
+    GameSetupView(viewModel: .init())
 }
